refactor(lazy-mesh): split phase3-config into per-domain spawn modules
`phase3-config.ts` was 1727 lines and 64 exports, the largest non-pinned SRP
offender in the repo. The 63 `phase3Add<Service>Mcp` functions now live in six
sibling modules - cloud (9), observability (5), security-quality (5), delivery
(12), data-bi (17), workplace (15) - over a shared helper module, leaving the
original file as the composition root plus re-exports. No consumer moved: the
only external import was `buildPhase3Servers`, and all 260 existing tests pass
unedited.

I15/D10 shaped the design. The rule accepts a file-local `wrap` alias ONLY in a
file that also declares it as a delegation to `wrapServerSpec` - deliberately, so
that naming a function `wrap` cannot launder an unwrapped spec. Each group module
therefore declares its own three-line delegation rather than importing a shared
one; an imported alias would leave every spawn site in the importing file
unrecognised and fail the rule closed on all of them. `phase3-shared.ts` declares
one too, since its `add*` helpers build specs. Both the shared module and the
rule's own comment now say so, because it reads like duplication worth removing
until you know why it is there.

Verified by breaking it, not by green: neutering one group's delegation to
`return spec` makes the auditor fire on every site in that file. I15 is live for
the new modules.

The split also surfaced rot in the I15 comment inventory. Nine rows cited
`phase3-config.ts:<line>`; SEVEN were already pointing at non-comment lines
before this change - `:424` read `sandboxCwd: string,` while claiming to attest
the Azure spawn site, which was at `:209`. `audit:doc-refs` checks only that a
cited line is in range, so the drift was invisible until moving the code pushed
the numbers past EOF. Following the precedent already set for the 27
per-connector I2 rows, the nine collapse into one row citing D10 - the rule that
enforces the property per site and fails the build - because a hand-maintained
list of line numbers cannot track a file that changes with every connector.

The rule's own doc comment was stale in the same way: it claimed 78 spawn sites
across three files. The real count at HEAD was 75. It now names the file shape
and no number.
